compositor|refactor: Decompose the representative sample builder
The representative sample plan is now composed from one builder per plan table, with its file bodies and its content-addressed blob store in modules of their own.

The plan fingerprint derives its source and target digests from the tables they cover rather than restating their ids beside them.
